寻迹与测距完成

+++ BSP/include/pwm_api.h
@@ -24,7 +24,7 @@ namespace bsp{
 
         void Stop();
 
-        void SetDuty(int duty);
+        void SetDuty(float duty);
 
         void SetPsc(int16_t psc);
